artifacts/telethon-digest/requirements.txt
telethon==1.36.0
aiohttp==3.9.5
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
httpx==0.27.2
